🚚(frontend) move styled components in appropriate folder
All reused styled components should be located in the folder 'components/Styled'
