+++ src/frontend/components/Loader/index.tsx
@@ -4,8 +4,8 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 import { v4 as uuidv4 } from 'uuid';
 
-import { theme } from '../../utils/theme/theme';
-import { Offscreen } from '../Offscreen';
+import { Offscreen } from 'components/Styled/Offscreen';
+import { theme } from 'utils/theme/theme';
 
 const Preloader = styled(Box)`
   width: 100%;
